add ber + snrdb (where available) to signal finder, allow to scroll main view again since it is longer than landscape mode allows
git-svn-id: svn://svn.freaque.net/enigma2@757 a295cab0-163a-0410-857a-8207b47ee9d6

+++ Classes/Objects/Generic/Signal.h
@@ -11,13 +11,13 @@
 @interface Signal : NSObject
 {
 @private
-	double snrdb;
+	float snrdb;
 	NSInteger snr;
 	NSInteger ber;
 	NSInteger agc;
 }
 
-@property (assign) double snrdb;
+@property (assign) float snrdb;
 @property (assign) NSInteger snr;
 @property (assign) NSInteger ber;
 @property (assign) NSInteger agc;
